+++ src/datasets.jl
@@ -226,7 +226,7 @@ end
 get_ndims_offset(f::JLDFile, dataspace::ReadDataspace, attributes::Nothing) =
     (dataspace.dimensionality, dataspace.dimensions_offset)
 
-function get_ndims_offset(f::JLDFile, dataspace::ReadDataspace, attributes::Vector{ReadAttribute})
+function get_ndims_offset(f::JLDFile, dataspace::ReadDataspace, attributes::AbstractVector)
     ndims = dataspace.dimensionality
     offset = dataspace.dimensions_offset
     if !isempty(attributes)
@@ -363,10 +363,9 @@ end
     psz += CONTINUATION_MSG_SIZE
 
     # Figure out the layout 
-    # The simplest CompactStorageMessage only supports data sets < 2^16
-    if datasz < 8192 || (!(data isa Array) && datasz < typemax(UInt16))
+    if datasz == 0 || (!(data isa Array) && datasz < 8192)
         layout_class = LcCompact
-        psz += jlsizeof(CompactStorageMessage) + datasz
+        psz += jlsizeof(Val(HmDataLayout); layout_class, data_size=datasz)
     elseif data isa Array && compress != false && isconcretetype(eltype(data)) && isbitstype(eltype(data))
         # Only now figure out if the compression argument is valid
         invoke_again, filter_id, compressor = get_compressor(compress)
@@ -377,7 +376,7 @@ end
         psz += chunked_storage_message_size(ndims(data)) + pipeline_message_size(filter_id::UInt16)
     else
         layout_class = LcContiguous
-        psz += jlsizeof(ContiguousStorageMessage)
+        psz += jlsizeof(Val(HmDataLayout); layout_class)
     end
 
     fullsz = jlsizeof(ObjectStart) + size_size(psz) + psz + 4
@@ -397,34 +396,33 @@ end
 
     # Data storage layout
     if layout_class == LcCompact
-        jlwrite(cio, CompactStorageMessage(datasz))
+        write_header_message(cio, Val(HmDataLayout); layout_class, data_size=datasz)
         if datasz != 0
             write_data(cio, f, data, odr, datamode(odr), wsession)
         end
-        jlwrite(cio, CONTINUATION_PLACEHOLDER)
+        write_continuation_placeholder(cio)
         jlwrite(io, end_checksum(cio))
     elseif layout_class == LcChunked
-    
         write_filter_pipeline_message(cio, filter_id)
 
-        # deflate first
         deflated = deflate_data(f, data, odr, wsession, compressor)
 
-        write_chunked_storage_message(cio, odr_sizeof(odr), size(data), length(deflated), h5offset(f, f.end_of_data))
-        
-        # Add NIL message replacable by continuation message
-        jlwrite(cio, CONTINUATION_PLACEHOLDER)
+        write_chunked_storage_message(cio, odr_sizeof(odr), size(data), length(deflated), h5offset(f, f.end_of_data))        
+        write_continuation_placeholder(cio)
         jlwrite(f.io, end_checksum(cio))
     
         seek(f.io, f.end_of_data)
         f.end_of_data += length(deflated)
         jlwrite(f.io, deflated)
     else
-        jlwrite(cio, ContiguousStorageMessage(datasz, h5offset(f, f.end_of_data)))
-        jlwrite(cio, CONTINUATION_PLACEHOLDER)
+        data_address = f.end_of_data + 8 - mod1(f.end_of_data, 8)
+        write_header_message(cio, Val(HmDataLayout); 
+            layout_class, data_address=h5offset(f, data_address), data_size=datasz)
+        write_continuation_placeholder(cio)
         jlwrite(io, end_checksum(cio))
 
-        f.end_of_data += datasz
+        f.end_of_data = data_address + datasz
+        seek(io, data_address)
         write_data(io, f, data, odr, datamode(odr), wsession)
     end
 
@@ -434,27 +432,15 @@ end
 function write_object_header_and_dataspace_message(cio::IO, f::JLDFile, psz::Int, dataspace::WriteDataspace)
     jlwrite(cio, ObjectStart(size_flag(psz)))
     write_size(cio, psz)
-
-    # Fill value
-    jlwrite(cio, HeaderMessage(HmFillValue, 2, 0))
-    jlwrite(cio, UInt8(3)) # Version
-    jlwrite(cio, 0x09)     # Flags
-
-    # Dataspace
-    jlwrite(cio, HeaderMessage(HmDataspace, jlsizeof(dataspace), 0))
-    jlwrite(cio, dataspace)
-
-    # Attributes
+    write_header_message(cio, Val(HmFillValue); flags=0x09)
+    write_header_message(cio, Val(HmDataspace); dataspace.dataspace_type, dimensions=dataspace.size)
     for attr in dataspace.attributes
-        jlwrite(cio, HeaderMessage(HmAttribute, jlsizeof(attr), 0))
-        write_attribute(cio, f, attr, f.datatype_wsession)
+        write_header_message(cio, f, attr)
     end
 end
 
-function write_datatype_message(cio::IO, datatype::H5Datatype)
-    jlwrite(cio, HeaderMessage(HmDatatype, jlsizeof(datatype), 1 | (2*isa(datatype, CommittedDatatype))))
-    jlwrite(cio, datatype)
-end
+write_datatype_message(cio::IO, dt::H5Datatype) = 
+    write_header_message(cio, Val(HmDatatype), 1 | (2*isa(dt, CommittedDatatype)); dt)
 
 
 @nospecializeinfer function write_dataset(f::JLDFile, @nospecialize(x), wsession::JLDWriteSession)::RelOffset
@@ -522,7 +508,7 @@ function delete_written_link!(f::JLDFile, roffset::RelOffset, name::AbstractStri
         if msg.type == HmLinkMessage && HmWrap(HmLinkMessage, msg).link_name == name
             # delete link
             seek(f.io, fileoffset(f, msg.offset))
-            jlwrite(f.io, HeaderMessage(HmNil, msg.size, 0))
+            write_header_message(f.io, Val(HmNil), 0, msg.size)
             update_checksum(f.io, iter.chunk.chunk_start, iter.chunk.chunk_end)
         end
     end